+++ geo-data-server/src/test/java/com/powsybl/geodata/server/GeoDataControllerTest.java
@@ -7,8 +7,6 @@
 package com.powsybl.geodata.server;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.github.nosan.embedded.cassandra.api.connection.CqlSessionCassandraConnection;
-import com.github.nosan.embedded.cassandra.spring.test.EmbeddedCassandra;
 import com.powsybl.geodata.extensions.Coordinate;
 import com.powsybl.geodata.server.dto.LineGeoData;
 import com.powsybl.geodata.server.dto.SubstationGeoData;
@@ -17,10 +15,8 @@ import com.powsybl.geodata.server.repositories.SubstationRepository;
 import com.powsybl.iidm.network.Country;
 import com.powsybl.iidm.network.test.EurostagTutorialExample1Factory;
 import com.powsybl.network.store.client.NetworkStoreService;
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.MockitoAnnotations;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
@@ -29,9 +25,6 @@ import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.UUID;
@@ -50,10 +43,9 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 @RunWith(SpringRunner.class)
 @WebMvcTest(GeoDataController.class)
 @ContextConfiguration(classes = {GeoDataApplication.class, CassandraConfig.class,
-        EmbeddedCassandraFactoryConfig.class, CqlCassandraConnectionFactoryTest.class})
-@EmbeddedCassandra(scripts = {"classpath:create_keyspace.cql", "classpath:geo_data.cql"})
+        EmbeddedCassandraFactoryConfig.class, CqlCassandraConnectionTestFactory.class})
 @DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_CLASS)
-public class GeoDataControllerTest {
+public class GeoDataControllerTest extends AbstractEmbeddedCassandraSetup  {
 
     @Autowired
     private ObjectMapper objectMapper;
@@ -69,27 +61,6 @@ public class GeoDataControllerTest {
 
     @MockBean
     private LineRepository lineRepository;
-
-    @Autowired
-    private CqlSessionCassandraConnection cqlSessionCassandraConnection;
-
-    @Before
-    public void setup() throws IOException {
-        MockitoAnnotations.initMocks(this);
-        String truncateScriptPath = getClass().getClassLoader().getResource("truncate.cql").getPath();
-        String truncateScript = Files.readString(Paths.get(truncateScriptPath));
-        executeScript(truncateScript);
-    }
-
-    public void executeScript(String script) {
-        String cleanedScript = script.replace("\n", "");
-        String[] requests = cleanedScript.split("(?<=;)");
-        for (String request : requests) {
-            if (!request.equals(" ")) {
-                cqlSessionCassandraConnection.execute(request);
-            }
-        }
-    }
 
     @Test
     public void test() throws Exception {
