+++ geo-data-server/src/test/java/com/powsybl/geodata/server/repositories/LineRepositoryTest.java
@@ -6,12 +6,10 @@
  */
 package com.powsybl.geodata.server.repositories;
 
-import com.github.nosan.embedded.cassandra.api.connection.CqlSessionCassandraConnection;
-import com.github.nosan.embedded.cassandra.spring.test.EmbeddedCassandra;
+import com.powsybl.geodata.server.AbstractEmbeddedCassandraSetup;
 import com.powsybl.geodata.server.CassandraConfig;
-import com.powsybl.geodata.server.CqlCassandraConnectionFactoryTest;
+import com.powsybl.geodata.server.CqlCassandraConnectionTestFactory;
 import com.powsybl.geodata.server.EmbeddedCassandraFactoryConfig;
-import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -19,9 +17,6 @@ import org.springframework.test.annotation.DirtiesContext;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringRunner;
 
-import java.io.IOException;
-import java.nio.file.Files;
-import java.nio.file.Paths;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -32,33 +27,12 @@ import static org.junit.Assert.assertFalse;
  * @author Chamseddine Benhamed <chamseddine.benhamed at rte-france.com>
  */
 @RunWith(SpringRunner.class)
-@ContextConfiguration(classes = {CassandraConfig.class, EmbeddedCassandraFactoryConfig.class, CqlCassandraConnectionFactoryTest.class})
-@EmbeddedCassandra(scripts = {"classpath:create_keyspace.cql", "classpath:geo_data.cql"})
+@ContextConfiguration(classes = {CassandraConfig.class, EmbeddedCassandraFactoryConfig.class, CqlCassandraConnectionTestFactory.class})
 @DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_CLASS)
-public class LineRepositoryTest {
+public class LineRepositoryTest extends AbstractEmbeddedCassandraSetup {
 
     @Autowired
     private LineRepository repository;
-
-    @Autowired
-    private CqlSessionCassandraConnection cqlSessionCassandraConnection;
-
-    @Before
-    public void setup() throws IOException {
-        String truncateScriptPath = getClass().getClassLoader().getResource("truncate.cql").getPath();
-        String truncateScript = Files.readString(Paths.get(truncateScriptPath));
-        executeScript(truncateScript);
-    }
-
-    public void executeScript(String script) {
-        String cleanedScript = script.replace("\n", "");
-        String[] requests = cleanedScript.split("(?<=;)");
-        for (String request : requests) {
-            if (!request.equals(" ")) {
-                cqlSessionCassandraConnection.execute(request);
-            }
-        }
-    }
 
     @Test
     public void test() {
